(PC-5221) refacto tab navigator and type helpers
