+++ src/features/navigation/__mocks__/RootNavigator.tsx
